+++ feature/main/src/main/java/kr/co/main/calendar/common/content/ScheduleContent.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import kr.co.main.calendar.common.CalendarCategoryIndicator
 import kr.co.main.model.calendar.ScheduleModel
 import kr.co.main.model.calendar.type.ScheduleModelType
-import kr.co.main.providers.FakeScheduleModelProvider
+import kr.co.main.providers.calendar.FakeScheduleModelProvider
 import kr.co.ui.theme.Paddings
 import kr.co.ui.theme.colors
 import kr.co.ui.theme.typo
